+++ packages/docs/reference/items/DataBind/stories/app.js
@@ -1,0 +1,4 @@
+import { registerComponents } from '@studiometa/js-toolkit';
+import { Action, DataBind, DataModel, DataScope } from '@studiometa/ui';
+
+registerComponents(Action, DataScope, DataBind, DataModel);