+++ packages/docs/reference/items/DataBind/stories/immediate.js
@@ -1,6 +1,4 @@
-import { registerComponent } from '@studiometa/js-toolkit';
+import { registerComponents } from '@studiometa/js-toolkit';
 import { DataBind, DataModel, DataScope } from '@studiometa/ui';
 
-registerComponent(DataScope);
-registerComponent(DataBind);
-registerComponent(DataModel);
+registerComponents(DataScope, DataBind, DataModel);